+++ tests/translation/dio/test_equality_between_types.py
@@ -1,6 +1,8 @@
 import pandas as pd
 import pytest
 
+from .conftest import TRANSLATED, UNTRANSLATED
+
 
 @pytest.mark.parametrize("ttype", [pd.DataFrame.from_dict, dict])
 def test_translate_and_insert(imdb_translator, ttype):
@@ -10,26 +12,3 @@ def test_translate_and_insert(imdb_translator, ttype):
         actual = actual.to_dict(orient="list")
 
     assert actual == TRANSLATED
-
-
-UNTRANSLATED = {
-    "firstTitle": [3, 41064, 15, 41068, 41069],
-    "nconst": [1, 15, 500, 30, 50],
-}
-
-TRANSLATED = {
-    "firstTitle": [
-        "3 not translated; default name=Title unknown",
-        "41064:Think Fast (original: Think Fast) from: 1949, to: 1950",
-        "15 not translated; default name=Title unknown",
-        "41068:Versatile Varieties (original: Versatile Varieties) from: 1949, to: 1951",
-        "41069:The Voice of Firestone (original: The Voice of Firestone) from: 1949, to: 1963",
-    ],
-    "nconst": [
-        "1:Fred Astaire from: 1899, to: 1987",
-        "15:James Dean from: 1931, to: 1955",
-        "500 not translated; default name=Name unknown",
-        "30:Audrey Hepburn from: 1929, to: 1993",
-        "50:Groucho Marx from: 1890, to: 1977",
-    ],
-}